+++ CHANGELOG.md
@@ -3,6 +3,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2026-04-09
+### Removed
+- deprecated `\Smsapi\Client\Feature\Contacts\Bag\CreateContactBag::withPhone`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsBag::setIdx`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\SendSmsToGroupBag::setIdx`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsToGroupBag::setIdx`
+- deprecated `\Smsapi\Client\Curl\Exception\RequestException`
+- deprecated `\Smsapi\Client\Curl\Exception\NetworkException`
+- deprecated `\Smsapi\Client\Curl\Exception\ClientException`
+
+## Changed
+- method signature `\Smsapi\Client\Feature\Contacts\ContactsFeature::findContacts`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsBag::setExternalId`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmssBag::setExternalId`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsToGroupBag::setExternalId`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\SendSmsBag::setExternalId`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\SendSmssBag::setExternalId`
+- method signature `\Smsapi\Client\Feature\Sms\Bag\SendSmsToGroupBag::setExternalId`
+
 ## [3.0.13] - 2026-01-12
 ### Added
 - PHP-8.5 support